Création commande (fix)
